Fixing a crash and a UI issue
PrefsByScreenUserControl does not have the correct size to show all the
buttons at the same time when inside PreferencesForm.
I am facing this issue in a high resolution screen in Windows 10.
I fixed it by making sure the sizes are exact.
I also made sure that the sub-form would stay in the same place when
switching tabs.